Sling type-safe model framework (#68)
Introduce a type-safe and editable framework for working with Sling resources. The out-of-the box sling models are a read-only solution. This framework will ensure node structures are maintained at the code level and hopefully minimize the amount of incorrect assumptions about the structure of content.

To create a new resource class simply extend SlingModel and add as many Field, DeepField, or ChildResource instances as desired to the class. These instances may be private or public, and while they don't need to be final it's best if they are.

Model objects need to have a one-arg constructor (the one arg is of Resource type) in this revision. This constructor is best left public.

See the SlingResourceTest class for examples.