create a single abstract class to encapsultate commonly used query logic (#97)
and in the process remove a lot of duplicate code.

The new AbstractJsonQueryServlet should serve as the super class for any servlet which needs to do a query on the server. It offers out-of-the-box pagination, custom queries and custom results.

This change also refactors two of the existing query servlets to use this new base class, thus eliminating a bunch of similar code, and hopefully simplifying the system going forward.