add a metadata extractor (#98)
The metadata extractor is an asciidoctor extension which is capable of extracting metadata from the Asciidoctor AST. In this revision only title and abstract are extracted, and with limitations. The metadata extractor is called after each module revision creation.

This change also fixes a memory leak in the Asciidoctor pool.