add an extension mechanism (#110)
This enables the possibility to add extension points to the system in asynchronous fashion. There are two main components in the mechanism:

The Events class serves as the single place where to fire an event. Events are hardcoded into methods and other system components should use this class to fire events when appropriate. On the receiving end, there are JobConsumers, which receive the events and process them. Events are placed in queues and JobConsumers are created for each one of these queues. There should be as many queues as JobConsumers are created. It is possible to create a single queue to process all events if desired.

The last layer is the EventProcessingExtension. This is an interface that should be implemented to process individual types of events. Aside from implementing this interface, extensions should also be declared as an OSGI component; this allows the system to pick them up from different bundles and process them. Each identified extension will be given the event to be processed. Failure to process the event will not stop other extensions from being invoked, and similarly will not revert the main operation which triggered the event.

The first (and only so far) event being fired is a post-publish event for a module revision. There are no extensions so far registered, but one possible extension would be to send the published module revision to a search index.