CCS-2965: add an action for unpublishing modules (#119)
This pull request sets the stage for module unpublishing. It offers an API to do an unpublish similar to the "publish" one. Once invoked, and only if there is a 'released' module revision, the system will unpublish the module and if there is no draft revision set, it will get the latest historical revision for the module and will make it the draft.

To invoke this API:

```
POST /module/path.adoc?:operation=pant:unpublish`
```