CCS-2975: replace module abstract extraction for an attribute-based approach (#131)
This change implements a more accurate way of identifying and extracting a module's abstract, relying on asciidoc's own syntax.

The approach requires the use of a single named attribute on any paragraph in the content to identify the abstract. The attribute's name is pantheon:abstract (clearly identifying the application and the purpose) to make sure writers are able to tell the platform where the metadata is without any room for ambiguity.

A sample usage for this would be:

= Why using attributes to identify metadata is a good idea

[pantheon:abstract]
It's a good idea to use attributes to identify metadata in the content as it removes any ambiguity when comparing to a positional based approach.