Introduce a slingstart project to provision a full sling application. (#159)
This PR changes the way the application is structured.

First it moves all the application code (backend and frontend) into a separate maven module called pantheon-bundle. This module will still allow for the individual deployment of the application bundle for development environments.

Another new maven module, pantheon-slingstart is added which takes care of packaging both sling and any additional bundles (in this revision just the aforementioned pantheon-bundle) into a single jar for full execution. This should simplify the packaging into a container image as a single jar will now be generated containing all necessary components to run the application. This will in turn pave the way for further modularization into more bundles.

Inside pantheon-slingstart there will be a jar file generated under the target directory. This jar may be run as the current sling file (java ... -jar pantheon-slingstart-xxx.jar) and it will run the full application.

Note: Slingstart will be deprecated in favor of the Sling features model, which might become part of the OSGI standard. When/if this happens the bundling work done in this PR will be helpful.