simplify the event handling mechanism (#169)
This change simplifies the event handling mechanism.

All events should now go through a single topic for processing. Each extension is responsible for filtering out the events for which they care.

All that is needed to create an extension point is to extend the EventProcessingExtension interface which has two methods: one to filter out all events that the extension doesn't care about, and another one to do the actual processing. The created class should be annotated with @Component(service = EventProcessingExtension.class)

The simplest filter can be by event type, as there are different types of events being fired.