bug fix: anonymous module rendering is not working (#182)
This change fixes a bug where anonymous users may not generate a module's preview. Since adding automatic metadata extraction from the asciidoc content, the process which writes the metadata to the JCR tree was not given the right permissions. This change ensures that the whole asciidoc generation process uses a service level resource resolver and thus ensure the appropriate write access.