CCS-3392: add logic to parse the module type from an asciidoc property (#224)
Pantheon will now detect the pantheon-module-type property declared inside a module's content to extract the module type. Valid values for this property are CONCEPT, PROCEDURE, and REFERENCE (all in caps). Any other value will be ignored by Pantheon.

The existing file-name-convention-based logic remains as a backup. This means that pantheon will look at the property first, and if not present, only then will it look at the file name to determine the module type.