eliminate the ResourceTraversal class (#393)
This change simplifies the mechanism for navigating through a model structure in a null-safe manner. It basically removes the ResourceTraversal class and instead uses the existing Child interface to accommodate the navigation functionality. It also removes the deprecated map method, which was the first attempt to offer a similar mechanism.

Most of the changes are about adapting to the new, simpler API. The core changes can be found in the following classes:
Child
Field
NullObjects
SlingModels