Update scripts (#490)
This change introduces an asynchronous, manual, and admin-only upgrade manager for the JCR tree.

It adds a framework for running system upgrades in a sequential and idempotent manner. Upgrades will need to be triggered manually by running the groovy script at /apps/pantheon/scripts/upgrade.groovy manually from a composum window.

The framework itself consists of three major components: an UpgradeRunner which is a system class that controls the execution and recording of any upgrades. An UpgradeProvider component interface which needs to be implemented by a component (a DefaultUpgradeProvider is included, with an empty list of upgrades) to offer an ordered list of Upgrade implementations (also not provided).

The system will join these three components and will run all the lised Upgrade implementations in order, committing changes to the JCR tree after each one is executed. If any given upgrade fails, the transaction will be rolled back, and the upgrade process will stop. The system will also keep a record of successfully executed upgrades, and will not execute them again; it will simply pick up from where it failed last time.